% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{aggregate_roc}
\alias{aggregate_roc}
\title{make aggregated roc curve}
\usage{
aggregate_roc(
  dat,
  subgroup.field = "Method",
  score = "BRCA1",
  lab = "fmut_brca1",
  mc.cores = 1,
  only_unique = F,
  include_group = FALSE,
  return_raw = FALSE
)
}
\arguments{
\item{dat}{data.table or data.frame of one row per observation. assumes each observation is a row}

\item{k}{number of groups}
}
\value{
a list of row ids corresponding to each fold and training and test split
}
\description{
make aggregated roc curve
}
\author{
Kevin Hadi
}
