% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_parse.R
\name{as_numeric_parse}
\alias{as_numeric_parse}
\title{Coerce to Numeric with Helpful Messages}
\usage{
as_numeric_parse(x, ...)
}
\arguments{
\item{x}{a string or vector to coerce to numeric}

\item{...}{additional arguments passed to `as.numeric()}
}
\description{
This function is a wrapper for as.numeric(). It takes a vector as input and
returns a numeric vector, while printing to the console every entry that
cannot be coerced.
}
\examples{
as_numeric_parse()
}
\keyword{KHtools}
