% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{asn2}
\alias{asn2}
\title{version of utils::assignInNamespace}
\usage{
asn2(x, value, ns, pos = -1, envir = as.environment(pos))
}
\description{
can be used to reassign function into a namespace
USE WITH CAUTION
}
