% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{behomology}
\alias{behomology}
\title{breakend exact homology}
\usage{
behomology(gg, hg, PAD = 2)
}
\arguments{
\item{gg}{gGraph (R6) object}

\item{hg}{character path to fasta or rtracklayer representation of fasta-like file}
}
\description{
pull out homology across exact breakends from gGraph object
}
