% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{binom.conf}
\alias{binom.conf}
\title{Get confidence intervals around fractions}
\usage{
binom.conf(n, tot, alpha = 0.025)
}
\value{
A vector
}
\description{
A convenience function to get confidence intervals around
proportions. To be used with gbar.error
}
