% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bool}
\alias{bool}
\title{Clean Up Boolean Logic}
\usage{
bool(x, nullignore = TRUE, na2false = TRUE)
}
\description{
wrapping around boolean statements to ignore NULL or length(0) vectors
in a series of boolean statements
}
