% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{bp2grl}
\alias{bp2grl}
\title{convert table of paired string coordinates to GRangesList}
\usage{
bp2grl(df, bp1.field = "bp1", bp2.field = "bp2", sort = TRUE)
}
\description{
convert table of paired string coordinates to GRangesList
}
\author{
Kevin Hadi
}
