% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{cenv}
\alias{cenv}
\title{concatenate environments}
\usage{
cenv(env = environment())
}
\arguments{
\item{env}{can be a data.frame/table, or list, or environment}
}
\description{
to be able to call a function within a function and access variables
use:
ev = "bla"
this.fun = function() anon()
datatable = data.table()
anon = function() {
    with(cenv(datatable), {print(ev)})
}
this.fun()
}
