% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{check_lst}
\alias{check_lst}
\title{check_lst
checking a list for any elements that are try-errors
usually from an lapply(..., function(x) try({})) call}
\usage{
check_lst(lst, class_condition = c("try-error", "error", "errored", "err"))
}
\arguments{
\item{lst}{A list}
}
\value{
a logical vector marking which elements are try-errors"
}
\description{
check_lst
checking a list for any elements that are try-errors
usually from an lapply(..., function(x) try({})) call
}
