% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{classystat}
\alias{classystat}
\title{calculates various scores from actual classes and predicted classes

Scores calculated from a classification task with known true labels:
precision, recall
F1, mean f1, weighted mean f1, mean precision,
mean recall, weighted mean recall, accuracy}
\usage{
classystat(real, pred)
}
\arguments{
\item{real}{true class labels}

\item{pred}{predicted class labels}
}
\value{
a list of various statistics for clasification task
}
\description{
calculates various scores from actual classes and predicted classes

Scores calculated from a classification task with known true labels:
precision, recall
F1, mean f1, weighted mean f1, mean precision,
mean recall, weighted mean recall, accuracy
}
