% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{coalesce}
\alias{coalesce}
\title{same as dplyr::coalesce, khtools:coalesce is an alias for khtools::clobber

clobber NA, or some value between multiple vectors
bads can be a function that returns a logical}
\usage{
coalesce(
  ...,
  bads = NA,
  bads.x = NULL,
  bads.y = NULL,
  r2l = FALSE,
  fromLast = FALSE,
  opposite = TRUE,
  comparefun = NULL
)
}
\description{
same as dplyr::coalesce, khtools:coalesce is an alias for khtools::clobber

clobber NA, or some value between multiple vectors
bads can be a function that returns a logical
}
