% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{colexists}
\alias{colexists}
\title{find out if column is in data.table}
\usage{
colexists(nm, df)
}
\arguments{
\item{nm}{name of column}

\item{df}{data.frame or data.table}
}
\description{
find out if column is in data.table
}
\author{
Kevin Hadi
}
