% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{colnames2<-}
\alias{colnames2<-}
\title{robust colnames() assignment}
\usage{
colnames2(x, useempty = FALSE) <- value
}
\arguments{
\item{x}{data with dimensions}
}
\value{
x data fully colnamed
}
\description{
robust colnames() assignment
}
