% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dcast.count2}
\alias{dcast.count2}
\title{Counts up occurrences from a melted table}
\usage{
\method{dcast}{count2}(
  tbl,
  lh,
  rh = NULL,
  countcol = "count",
  wt = 1,
  fun.aggregate = "sum",
  value.var = "dummy",
  ...
)
}
\value{
A data frame or data.table
}
\description{
Counting up occurrences in a table while taking factor levels into account
Also allows for weighting the counts using flexible argument parsing
Can either provide a weight as a name of a column,
as values themselves, or don't provide at all, and the function looks for a
column named "wt" for its values
}
