% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dcast.wrap}
\alias{dcast.wrap}
\title{wrapper around dcast or dcast2

A convenience wrapper around dcast to make formula generation more
programmatic}
\usage{
dcast.wrap(x, lh, rh, dcast.fun, ...)
}
\value{
A data frame or data.table
}
\description{
wrapper around dcast or dcast2

A convenience wrapper around dcast to make formula generation more
programmatic
}
