% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{debug.s4}
\alias{debug.s4}
\title{debug an S4 function

wrapper around trace}
\usage{
debug.s4(what, signature, where)

debug.s4(what, signature, where)
}
\description{
Debugging an S4 function can't be done with debug().
This function is a convenience wrapper around trace()
to step into S4 methods
}
