% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dedup.cols}
\alias{dedup.cols}
\title{applies dedup to colnames}
\usage{
dedup.cols(tbl, remove = FALSE)
}
\value{
A data.table or data.frame
}
\description{
dedup the column names of a data.frame/data.table
}
