% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dedup}
\alias{dedup}
\title{dedup}
\usage{
dedup(x, suffix = ".")
}
\arguments{
\item{x}{vector to dedup}

\item{suffix}{character separator}
}
\value{
a vector
}
\description{
stolen from skitools
}
\author{
Marcin Imielinski
}
