% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{df2gr}
\alias{df2gr}
\title{data frame to GRanges}
\usage{
df2gr(
  df,
  seqnames.field = "seqnames",
  start.field = "start",
  end.field = "end",
  strand.field = "strand",
  ignore.strand = FALSE,
  keep.extra.columns = TRUE
)
}
\value{
GRanges
}
\description{
data frame to GRanges
}
\author{
Kevin Hadi
}
