% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dg}
\alias{dg}
\title{alias of dynget}
\usage{
dg(
  x,
  px = TRUE,
  ifnotfound = stop(gettextf("\%s not found", sQuote(x)), domain = NA),
  minframe = 0L,
  inherits = FALSE
)
}
\description{
convenience wrapper around dynget
}
