% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dig_dir2}
\alias{dig_dir2}
\title{dig into a file path's directory

Convenience wrapper around dir2() to pull out files from the same
directory of a given file.}
\usage{
dig_dir2(
  x,
  pattern = NULL,
  full.names = TRUE,
  mc.cores = 1,
  unlist = TRUE,
  do_dirname = TRUE,
  ...
)
}
\description{
dig into a file path's directory

Convenience wrapper around dir2() to pull out files from the same
directory of a given file.
}
\author{
Kevin Hadi
}
