% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{diginjob}
\alias{diginjob}
\title{Dig into inputs Flow Job that generated an output}
\usage{
diginjob(x)
}
\value{
data.table
}
\description{
takes a path of an output of a flow job and looks for
Job.rds and reads in input
}
