% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{digjob}
\alias{digjob}
\title{Dig into Flow Job that generated an output}
\usage{
digjob(x, readin = TRUE, get_inputs = FALSE)
}
\value{
Job
}
\description{
takes a path of an output of a flow job and looks for
Job.rds and reads in Job object
}
