% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dir2}
\alias{dir2}
\title{dir with full grep}
\usage{
dir2(
  path = ".",
  pattern = NULL,
  all.files = FALSE,
  full.names = FALSE,
  recursive = FALSE,
  ignore.case = FALSE,
  include.dirs = FALSE,
  no.. = FALSE,
  ...
)
}
\description{
dir with full grep
}
\author{
Kevin Hadi
}
