% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{do.cols}
\alias{do.cols}
\title{do.cols}
\usage{
do.cols(x, rmcol = 1, FUN = rowSums, by.FUN = NULL, exact = F, invert = T)
}
\arguments{
\item{x}{object with ncol >= 1}
}
\description{
do.cols
}
\author{
Kevin Hadi
}
