% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_utils.R
\name{do.pp}
\alias{do.pp}
\title{do caret preProcess}
\usage{
do.pp(
  dat,
  vars,
  method = c("center", "scale"),
  prefun = NULL,
  rangeBounds = c(0, 1),
  verbose = TRUE
)
}
\description{
do caret preProcess
}
