% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_any2lg}
\alias{dt_any2lg}
\title{convert all columns to logical

set columns to logical}
\usage{
dt_any2lg(dt, cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}
}
\value{
A data.table
}
\description{
convert all columns to logical

set columns to logical
}
