% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_empty2na}
\alias{dt_empty2na}
\title{convert columns with empty character to NA

coerce "" in columns of class "character" to NA_character_}
\usage{
dt_empty2na(dt)
}
\arguments{
\item{dt}{data.table}

\item{these_cols}{NULL by default, will select columns of character class, otherwise will be specified}
}
\value{
A data.table
}
\description{
convert columns with empty character to NA

coerce "" in columns of class "character" to NA_character_
}
