% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_f2char}
\alias{dt_f2char}
\title{convert factor columns to character

coerce factor columns to character}
\usage{
dt_f2char(dt, cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}
}
\value{
A data.table
}
\description{
convert factor columns to character

coerce factor columns to character
}
