% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_lg2int}
\alias{dt_lg2int}
\title{logical to integer in data tables

coerce logical columns to integers in data table}
\usage{
dt_lg2int(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
A data.table
}
\description{
logical to integer in data tables

coerce logical columns to integers in data table
}
