% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_na2true}
\alias{dt_na2true}
\title{convert columns with NA to true

coerce NA in columns of class "logical" to TRUE}
\usage{
dt_na2true(dt, these_cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{these_cols}{NULL by default, will select columns of class logical, otherwise will be specified}
}
\value{
A data.table
}
\description{
convert columns with NA to true

coerce NA in columns of class "logical" to TRUE
}
