% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_na2zero}
\alias{dt_na2zero}
\title{convert columns with NA to zeros

coerce NA in columns of class "numeric"/"integer" to FALSE}
\usage{
dt_na2zero(dt, these_cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{these_cols}{NULL by default, will select columns of numeric class, otherwise will be specified}
}
\value{
A data.table
}
\description{
convert columns with NA to zeros

coerce NA in columns of class "numeric"/"integer" to FALSE
}
