% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_setallna}
\alias{dt_setallna}
\title{convert all data.table entries to NA

Set all columns to NA...
can't remember why one would want this...}
\usage{
dt_setallna(dt, cols = NULL, na_type = NA_integer_)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}
}
\value{
A data.table
}
\description{
convert all data.table entries to NA

Set all columns to NA...
can't remember why one would want this...
}
