% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_setchar}
\alias{dt_setchar}
\title{convert column to character

set columns to character
default is nonsensical?}
\usage{
dt_setchar(dt, cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}
}
\value{
A data.table
}
\description{
convert column to character

set columns to character
default is nonsensical?
}
