% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_setint}
\alias{dt_setint}
\title{convert column to integer}
\usage{
dt_setint(dt, cols = NULL)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}
}
\value{
A data.table
}
\description{
coerce column of type "numeric" by default to "integer"
}
