% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_setnull}
\alias{dt_setnull}
\title{set columnn to NULL in data.table

remove column from data table by setting to NULL}
\usage{
dt_setnull(dt, cols)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}
}
\value{
A data.table
}
\description{
set columnn to NULL in data.table

remove column from data table by setting to NULL
}
