% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dtapply}
\alias{dtapply}
\title{mclapply on a table split by a column}
\usage{
dtapply(
  tbl,
  split_col = "system_id",
  FUN,
  mc.cores = 1,
  mc.strict = TRUE,
  split_col_sort = FALSE,
  mclapply = parallel::mclapply,
  ...
)
}
\description{
Do function on splits of a table in parallel
}
