% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_write.R
\name{duckdb_create_and_replace_table}
\alias{duckdb_create_and_replace_table}
\title{duckdb_create_and_replace_table}
\usage{
duckdb_create_and_replace_table(con, target, select_sql)
}
\arguments{
\item{con}{db connection}

\item{target}{navn på måltabell}

\item{select_sql}{uttrykk som genererer den nye tabellen}
}
\description{
Erstatter CREATE OR REPLACE X AS SELECT FROM X, ved at det først
skrives en tmp_tabell, som deretter overskriver måltabellen.
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
