% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_write.R
\name{duckdb_drop_tables}
\alias{duckdb_drop_tables}
\title{duckdb_drop_tables}
\usage{
duckdb_drop_tables(con, tables)
}
\arguments{
\item{con}{db connection}

\item{tables}{liste over tabeller som skal droppes}
}
\description{
drops tables if they exist
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_and_replace_table]{duckdb_create_and_replace_table()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
