% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_alter.R, R/duckdb_utils.R
\name{duckdb_ensure_columns}
\alias{duckdb_ensure_columns}
\title{duckdb_ensure_columns}
\usage{
duckdb_ensure_columns(con, table, cols)

duckdb_ensure_columns(con, table, cols)
}
\arguments{
\item{con}{db connection}

\item{table}{table name}

\item{cols}{navngitt vektor eller liste på formatet c(KOL = "TYPE", KOL2 = "TYPE") eller list(KOL = ...)}
}
\description{
Sikrer at kolonner eksisterer, eller oppretter dem som tomme kolonner med riktig type.

Sikrer at kolonner eksisterer, eller oppretter dem som tomme kolonner med riktig type.
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_and_replace_table]{duckdb_create_and_replace_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
