% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_init.R
\name{duckdb_init}
\alias{duckdb_init}
\title{duckdb_init}
\usage{
duckdb_init(dbname = NULL, mem_limit_gb = 8)
}
\arguments{
\item{dbname}{name of the duckdb-file}

\item{mem_limit_gb}{grense for hvor mange GB RAM databasen får lov å benytte}
}
\value{
en databaseconnection
}
\description{
Oppretter en arbeidsdatabase med en satt grense for RAM-bruk og sørger for at denne er tom
Oppretter også en temp-database for eventuelt spilling til disk
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_and_replace_table]{duckdb_create_and_replace_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
