% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dunlist2}
\alias{dunlist2}
\title{unlist into a data.table

unlisting a list into a data table with
ids corresponding to each list element}
\usage{
dunlist2(x, simple = FALSE)
}
\description{
unlist into a data.table

unlisting a list into a data table with
ids corresponding to each list element
}
