% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dynget}
\alias{dynget}
\title{modification of base::dynGet()

slight modification of base::dynGet()
minframe set to 0 to also look in global environment
and it's robust to using within functions}
\usage{
dynget(
  x,
  px = TRUE,
  ifnotfound = stop(gettextf("\%s not found", sQuote(x)), domain = NA),
  minframe = 0L,
  inherits = FALSE
)
}
\description{
modification of base::dynGet()

slight modification of base::dynGet()
minframe set to 0 to also look in global environment
and it's robust to using within functions
}
