% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{enframe}
\alias{enframe}
\title{same as tibble::enframe}
\usage{
enframe(x, name = "name", value = "value", as.data.table = TRUE)
}
\arguments{
\item{x}{named vector}

\item{name}{name of column containing names(x)}

\item{value}{name of column containing x values}
}
\description{
kevin's implementation of tibble::enframe()
}
