% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{enframe_list}
\alias{enframe_list}
\title{data table-ize list elements and add name column}
\usage{
enframe_list(
  lst,
  name = "name",
  value = "value",
  as.data.table = TRUE,
  rbind = TRUE,
  mc.cores = 1
)
}
\description{
data table-ize list elements and add name column
}
