% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{est_snv_cn}
\alias{est_snv_cn}
\title{estimate snv cn}
\usage{
est_snv_cn(gr, jabba, somatic = FALSE)
}
\arguments{
\item{gr}{GRanges}

\item{jab}{jabba rds or jabba list object}
}
\description{
estimate snv cn
}
