% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{et}
\alias{et}
\title{shortcut for eval(parse(text = <string>))}
\usage{
et(txt, eval = TRUE, envir = parent.frame())
}
\arguments{
\item{txt}{string to evaluate}
}
\description{
shortcut for eval(parse(text = <string>))
}
\author{
Kevin Hadi
}
