% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{extract_ggplot}
\alias{extract_ggplot}
\title{helper function to grab relevant ggplot_build outputs}
\usage{
extract_ggplot(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{A ggplot object}
}
\description{
takes an input function and grabs the processed data frame
that ggplot employs to draw its plots

Note: reordering plot data to fit the ggplot wrangled data is
(if there is a 1 to 1 mapping)
plot_data[order(<facet.x>, <facet.y>, <grouping_variable>)]

OR

plot_data[order(PANEL, <grouping_variable>)]

useful for custom construction of layers
}
