% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ez_string}
\alias{ez_string}
\title{don't use this. use base::dput()}
\usage{
ez_string(string_vec, c = T, list = !c, quotes = T, ws = "\\n")
}
\description{
Don't use this... use base::dput() An easy way to print out a c() vector into a blog file
}
