% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{false2na}
\alias{false2na}
\title{replace FALSE with NA}
\usage{
false2na(x)
}
\value{
A logical vector with NAs set to FALSE
}
\description{
A convenience function to set a logical vector with NAs to false
}
