% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_utils.R
\name{feature_importance}
\alias{feature_importance}
\title{feature_importance}
\usage{
feature_importance(
  ixfold,
  xfolds,
  dat,
  lambda = 0,
  vars,
  seed = 10,
  ycol,
  debug = FALSE,
  rpart = FALSE,
  permute_varlist = NULL,
  added_testd = NULL,
  caret_preprocess = do.pp(traind, vars = vars, method = "range", prefun = log1p),
  only_positive = FALSE,
  s = 0,
  alpha = 1
)
}
\description{
feature_importance
}
