% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{file.exists2}
\alias{file.exists2}
\title{slightly more robust test for whether file exists}
\usage{
file.exists2(x, nullfile = "/dev/null", bad = c(NA, "NA", "NULL", ""))
}
\arguments{
\item{x}{a character vector}
}
\value{
logical
}
\description{
test whether file exists. "/dev/null", NA, "NA", "NULL", values excluded
by default.
}
\author{
Kevin Hadi
}
