% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{file.info2}
\alias{file.info2}
\title{file.info2}
\usage{
file.info2(fn, col = NULL, include.all = FALSE)
}
\value{
data.frame/data.table
}
\description{
A more robust file.info2 that removes any paths that do not exist
}
