% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{file.not.exists}
\alias{file.not.exists}
\title{slightly more robust test for whether file does not exist

test whether a file is NA, NULL, or /dev/null OR if 
the file exists}
\usage{
file.not.exists(x, nullfile = "/dev/null", bad = c(NA, "NA", "NULL", ""))
}
\arguments{
\item{x}{a character vector}
}
\value{
logical
}
\description{
slightly more robust test for whether file does not exist

test whether a file is NA, NULL, or /dev/null OR if 
the file exists
}
\author{
Kevin Hadi
}
