% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{fillby}
\alias{fillby}
\title{fill in variables of data table by combos}
\usage{
fillby(x, by, fillcol, fill = 0L, use_factor_levels = TRUE)
}
\description{
especially useful for expanding factors excluded from a query
}
