% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{find_dups}
\alias{find_dups}
\title{find all duplicates in a vector}
\usage{
find_dups(..., re_sort = FALSE, sep = " ", as.logical = FALSE)
}
\arguments{
\item{vec}{A vector}
}
\value{
a logical vector with all positions marked TRUE being duplicates
}
\description{
find all elements that have duplicates in a vector
}
\examples{
find_dups(c(1,1,1,3,5))
find_dups(c(1,3,1,3,1))
find_dups(c(3,1,5,4,4))
}
