% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{fit.cnv.sig}
\alias{fit.cnv.sig}
\title{fit battenberg copy number to CN signatures (Nature 2022)}
\usage{
fit.cnv.sig(
  gr.seg,
  sig.cnv = "~/Dropbox/Isabl/HRD/Steele-cnv-signature-definitions.txt",
  id = NULL
)
}
\description{
fit battenberg copy number to CN signatures (Nature 2022)
}
