% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_utils.R
\name{fitcvglmnet}
\alias{fitcvglmnet}
\title{fit cv.glmnet}
\usage{
fitcvglmnet(
  var,
  lambda = mylambdas(),
  family = "multinomial",
  ycol,
  dat,
  only_positive = TRUE,
  nlambda = 100,
  maxit = 100000,
  thresh = 0.0000001,
  type.multinomial = "grouped",
  standardize = FALSE,
  standardize.response = FALSE,
  type.measure = "default",
  alpha = 1,
  ...
)
}
\description{
fit cv.glmnet
}
