% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{fix.cols}
\alias{fix.cols}
\title{Fix messed up data frame/table column names}
\usage{
fix.cols(dt, sep = "_")
}
\arguments{
\item{dt}{data.table/frame}

\item{sep}{separator field}
}
\value{
data.table/frame
}
\description{
If there are any malformed columns 
(e.g. those with numbers at the beginning
or a dash) these column names are fixed
}
