% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{forcefun}
\alias{forcefun}
\title{force functions to load}
\usage{
forcefun(envir = globalenv(), evalenvir = globalenv())
}
\arguments{
\item{envir}{environment to grab all functions from}

\item{evalenvir}{environment to evaluate in (probably doesn't matter)}
}
\description{
A function to evaluate all functions in a single environment
}
