% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{forceload}
\alias{forceload}
\title{force functions to load from all libraries}
\usage{
forceload(envir = globalenv())
}
\arguments{
\item{envir}{environment to evaluate in (probably doesn't matter)}
}
\description{
A function to evaluate all functions in all loaded
and attached packages to prevent errors upon reinstallation of
a package
}
