% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{getcache}
\alias{getcache}
\title{getcache

Get the path of a Flow job object's cache.}
\usage{
getcache(object)
}
\value{
A character
}
\description{
getcache

Get the path of a Flow job object's cache.
}
