% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{getdat}
\alias{getdat}
\title{a method to get the "data" argument from a with/within expression

to be used within "with()" within the expression}
\usage{
getdat(n = 0L)
}
\value{
data.frame/data.table
}
\description{
a method to get the "data" argument from a with/within expression

to be used within "with()" within the expression
}
