% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{getdat2}
\alias{getdat2}
\title{getdat2}
\usage{
getdat2(nm = "data")
}
\description{
another function to use inside the expression argument of "with/within" family
to grab the enclosing data environment
}
