% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gg.hist}
\alias{gg.hist}
\title{generate ggplot histogram}
\usage{
gg.hist(
  dat.x,
  as.frac = FALSE,
  bins = 50,
  center = NULL,
  boundary = NULL,
  trans = "identity",
  print = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = "",
  x_breaks = 20,
  y_breaks = 10,
  expand = waiver(),
  ...
)
}
\value{
A ggplot object
}
\description{
generate ggplot histogram
}
