% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gg.sline}
\alias{gg.sline}
\title{wrapper around geom_point and geom_smooth

a wrapper around geom_smooth to fit dot plot with lm line}
\usage{
gg.sline(
  x,
  y,
  group = "x",
  colour = NULL,
  smethod = "lm",
  dens_type = c("point", "hex"),
  facet1 = NULL,
  facet2 = NULL,
  transpose = FALSE,
  facet_scales = "fixed",
  line = TRUE,
  formula = y ~ x,
  print = FALSE,
  hex_par = list(bins = 50),
  wes = NULL,
  cex.scatter = 0.1
)
}
\arguments{
\item{x}{numeric values}

\item{y}{numeric values}

\item{group}{vector parallel to x and y that specifies grouping}

\item{colour}{vector that specifies colouring of points}
}
\value{
A ggplot object
}
\description{
wrapper around geom_point and geom_smooth

a wrapper around geom_smooth to fit dot plot with lm line
}
