% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gg_mytheme}
\alias{gg_mytheme}
\title{custom theme for ggplot

A custom theme for ggplots}
\usage{
gg_mytheme(
  gg,
  base_size = 16,
  legend.position = "none",
  flip_x = TRUE,
  x_axis_cex = 1,
  y_axis_cex = 1,
  ylab_cex = 1,
  xlab_cex = 1,
  title_cex = 1,
  x_angle = 90,
  x_axis_hjust = 1,
  x_axis_vjust = 0.5,
  y_axis_hjust = 1,
  print = FALSE
)
}
\value{
A ggplot object
}
\description{
custom theme for ggplot

A custom theme for ggplots
}
