% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ggmroc}
\alias{ggmroc}
\title{ggplot for multiROC

helper function for outputting ggplot}
\usage{
ggmroc(gdat, palette = "Moonrise2")
}
\arguments{
\item{lbl}{output from mroclab}

\item{prd}{output from mrocpred}
}
\value{
A data.frame that can be fed into ggplot
}
\description{
ggplot for multiROC

helper function for outputting ggplot
}
