% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{glm.nb2}
\alias{glm.nb2}
\title{glm.nb2

Run a negative binomial regression.
If it fails, run a poisson regression}
\usage{
glm.nb2(...)
}
\value{
A GLM model
}
\description{
glm.nb2

Run a negative binomial regression.
If it fails, run a poisson regression
}
