% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{globasn}
\alias{globasn}
\title{assign an object to global environment}
\usage{
globasn(obj, var = NULL, return_obj = TRUE, envir = .GlobalEnv,
  verbose = TRUE)
}
\arguments{
\item{obj}{The object to assign to the global environment}

\item{var}{Optional name of variable, specified as string}
}
\value{
either NULL or the object being assigned
}
\description{
ONLY USE IF YOU KNOW WHAT YOU ARE DOING
This function forces assignment of a variable/function
to the global environment
}
