% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{good.file}
\alias{good.file}
\title{Does File exists and is file size greater than threshold

Queries a set of file paths for whether the file exists AND
if the file is greater than a size threshold}
\usage{
good.file(x, size.thresh = 0)
}
\arguments{
\item{x}{character vector of file paths}

\item{size.thresh}{threshold of minimum file size}
}
\value{
logical
}
\description{
Does File exists and is file size greater than threshold

Queries a set of file paths for whether the file exists AND
if the file is greater than a size threshold
}
