% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.disjoin}
\alias{gr.disjoin}
\title{updated gr.disjoin from gUtils to work with GRangesList}
\usage{
gr.disjoin(x, ..., ignore.strand = TRUE)
}
\value{
GRanges or GRangesList
}
\description{
updated gr.disjoin from gUtils to work with GRangesList
}
\author{
Kevin Hadi
}
