% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.fixseq}
\alias{gr.fixseq}
\title{get permissive seqlengths from multiple GRanges-like objects}
\usage{
gr.fixseq(...)
}
\value{
GRanges
}
\description{
takes seqlengths
May need to set either of these via
Sys.setenv(DEFAULT_BSGENOME = "path_to_ref.chrom.sizes")
Sys.setenv(DEFAULT_GENOME = "path_to_ref.chrom.sizes")
}
\author{
Kevin Hadi
}
