% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.flipstrand}
\alias{gr.flipstrand}
\title{works with GRangesLists}
\usage{
gr.flipstrand(gr)
}
\value{
A GRanges with the by metadata field attached to the seqnames
}
\description{
to be used with gr_construct_by
}
\author{
Kevin Hadi
}
