% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.order}
\alias{gr.order}
\title{order granges, grangeslist}
\usage{
gr.order(gr, ignore.strand = T)
}
\value{
GRanges
}
\description{
order granges or grangeslist by seqlevels
also reorders seqlevels into 1:22, X, Y format
}
\author{
Kevin Hadi
}
