% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.patch}
\alias{gr.patch}
\title{same as gr.fix basically}
\usage{
gr.patch(gr, patch, onlystandard = TRUE)
}
\value{
GRanges
}
\description{
need to change implementation...
it doesn't add to gr.fix so far
except to add default genome seqlengths
on top of whatever is already part of the input
}
\author{
Kevin Hadi
}
