% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.round}
\alias{gr.round}
\title{round window to nearest unit}
\usage{
gr.round(gr, nearest = 10000, all = TRUE, reduce = FALSE)
}
\value{
GRanges
}
\description{
Rounding window to nearest unit.
meant to be used for plotting window in gTrack.
For that purpose, it's most useful to use reduce = TRUE
}
\author{
Kevin Hadi
}
