% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.split}
\alias{gr.split}
\title{split a gr by field(s) in elementMetadata of GRanges or a given vector}
\usage{
gr.split(gr, ..., sep = " ")
}
\value{
GRanges
}
\description{
split GRanges by field(s)
if providing a variable not already within the GRanges,
may need to use dynget(variable_name)
}
\author{
Kevin Hadi
}
