% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.sprange}
\alias{gr.sprange}
\title{get range based on a field(s) to split by in elementMetadata of GRanges, or given vector}
\usage{
gr.sprange(
  gr,
  ...,
  ignore.strand = FALSE,
  pad = 0,
  return.grl = FALSE,
  sep = paste0(" ", rand.string(length = 8), " ")
)
}
\value{
GRanges
}
\description{
split and get range of GRanges by field(s)
if providing a variable not already within the GRanges,
may need to use dynget(variable_name)
}
\author{
Kevin Hadi
}
