% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.spreduce}
\alias{gr.spreduce}
\title{reduce based on a field(s) to split by in elementMetadata of GRanges, or given vector}
\usage{
gr.spreduce(gr, ..., pad = 0, sep = " ")
}
\value{
GRanges
}
\description{
split and reduce GRanges by field(s)
if providing a variable not already within the GRanges,
may need to use dynget(variable_name)
}
\author{
Kevin Hadi
}
