% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.strand}
\alias{gr.strand}
\title{specify strand of granges or grangeslist}
\usage{
gr.strand(gr, str = "*")
}
\value{
granges or grangeslist
}
\description{
specify strand of granges or grangeslist
}
\author{
Kevin Hadi
}
