% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.sub2}
\alias{gr.sub2}
\title{Substitute chr in GRanges or GRangesList}
\usage{
gr.sub2(x, sub.pt = c("^chr", "MT"), sub.rp = c("", "M"))
}
\value{
GRanges or GRangesList
}
\description{
substitute characters in seqnames
}
\author{
Kevin Hadi
}
