% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr2df}
\alias{gr2df}
\title{granges to datatable via dataframe}
\usage{
gr2df(gr, var = "rowname", as.data.table = FALSE)
}
\value{
GRangesList
}
\description{
granges to datatable via dataframe
}
\author{
Kevin Hadi
}
