% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr_calc_cov}
\alias{gr_calc_cov}
\title{output data structure for ski slope from anchorlifted SNV}
\usage{
gr_calc_cov(
  gr,
  PAD = 50,
  field = NULL,
  start.base = -1e+06,
  end.base = -5000,
  win = 10000,
  FUN = "mean",
  baseline = NULL,
  normfun = "*",
  normfactor = NULL
)
}
\value{
A GRanges
}
\description{
output data structure for ski slope from anchorlifted SNV
}
