% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{grab.hrdetect.features}
\alias{grab.hrdetect.features}
\title{grab hrdetect features from hrdetect results}
\usage{
grab.hrdetect.features(hrdetect_results, goodpairs, field, id.field = id.field)
}
\description{
read in HRDetect results
}
