% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{grabcov}
\alias{grabcov}
\title{convenience function to pull out coverage data from table}
\usage{
grabcov(
  pairs,
  id = NULL,
  field = "decomposed_cov",
  y.field = NULL,
  rel2abs = T
)
}
\value{
GRanges of coverage data
}
\description{
convenience function to pull out coverage data from table
}
