% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{grabggtrack}
\alias{grabggtrack}
\title{convenience function to pull out jabba model + coverage}
\usage{
grabggtrack(
  pairs,
  pair = NULL,
  jab_field = "complex",
  cov_field = "decomposed_cov",
  cov_y_field = NULL
)
}
\description{
convenience function to pull out jabba model + coverage
}
