% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{\%=\%}
\alias{\%=\%}
\title{test if two vectors are equal (uses conversion to character)

Robust to NA}
\usage{
x \%=\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
logical vector
}
\description{
test if two vectors are equal (uses conversion to character)

Robust to NA
}
\author{
Kevin Hadi
}
