% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{\%k\%}
\alias{\%k\%}
\title{similar to setkey except a general use utility

slower version of setkey, but for interactive use}
\usage{
x \%k\% y
}
\description{
similar to setkey except a general use utility

slower version of setkey, but for interactive use
}
