% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{\%Q\%.CompressedGRangesList}
\alias{\%Q\%.CompressedGRangesList}
\title{query on CompressedGRangesList}
\usage{
\method{\%Q\%}{CompressedGRangesList}(x, y)
}
\value{
GRangesList
}
\description{

}
\author{
Kevin Hadi
}
