% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{grep_col_sort}
\alias{grep_col_sort}
\title{grep_col_sort}
\usage{
grep_col_sort(
  patterns,
  df,
  all_cols = TRUE,
  match_first = TRUE,
  perl = TRUE,
  fixed = FALSE
)
}
\value{
A data.frame/data.table
}
\description{
order columns of a data.frame/data.table based on the supplied
order of multiple character patterns.
}
