% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{grep_order}
\alias{grep_order}
\title{grep_order}
\usage{
grep_order(
  patterns,
  text,
  return_na = FALSE,
  first_only = FALSE,
  perl = TRUE,
  fixed = FALSE
)
}
\value{
A character vector
}
\description{
order text based on the supplied order of multiple patterns
}
