% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gaps}
\alias{gaps}
\alias{gaps,GRangesList-method}
\title{gaps on GRangesList}
\usage{
\S4method{gaps}{CompressedGRangesList}(x, start = 1L, end = seqlengths(x))
}
\value{
GRangesList
}
\description{

}
\author{
Kevin Hadi
}
