% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{grok_vcf}
\alias{grok_vcf}
\title{modded grok_vcf}
\usage{
grok_vcf(
  x,
  label = NA,
  keep.modifier = TRUE,
  long = FALSE,
  oneliner = FALSE,
  verbose = FALSE,
  geno = NULL,
  tmp.dir = tempdir(),
  gr = NULL
)
}
\arguments{
\item{x}{path to vcf}
}
\value{
GRanges
}
\description{
modded grok_vcf
}
\author{
Marcin Imielinski
}
