% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gt.each}
\alias{gt.each}
\title{convenience function to plot each element separately

useful for grangeslists and gwalks to plot each one on a separate track
good for visualization}
\usage{
gt.each(gr)
}
\value{
gTrack
}
\description{
convenience function to plot each element separately

useful for grangeslists and gwalks to plot each one on a separate track
good for visualization
}
