% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gt.fix}
\alias{gt.fix}
\title{fix gtrack metadata elements for plotting}
\usage{
gt.fix(gt, lwd.scale = 1, lwd.border.scale = 1, ywid.scale = 1)
}
\value{
A data.frame/data.table
}
\description{
Parse a tabular summary of a glm model
}
