% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{header}
\alias{header}
\title{header}
\usage{
header(..., .silent = getOption("khtools.silent", FALSE))
}
\arguments{
\item{...}{objekter som skal skrives ut}

\item{.silent}{mulighet til å skru av logging ved testing}
}
\description{
Skriver overskrift til konsoll og flusher output.
}
\seealso{
Other utilities:
\code{\link[=msg]{msg()}}
}
\concept{utilities}
