% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{integer_breaks}
\alias{integer_breaks}
\title{make integer breaks on axes

a la scales::breaks_pretty}
\usage{
integer_breaks(n = 5, ...)
}
\description{
make integer breaks on axes

a la scales::breaks_pretty
}
