% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{interaction2}
\alias{interaction2}
\title{interaction but orders levels based on input vectors}
\usage{
interaction2(..., drop = FALSE, sep = ".", lex.order = FALSE)
}
\description{
Same as base::interaction but orders levels based on the appearance of elements
in input vector(s)
}
