% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{intercalate}
\alias{intercalate}
\title{collate two vectors together}
\usage{
intercalate(...)
}
\arguments{
\item{...}{A set of vectors to collate}
}
\value{
a vector with values of inputs collated together
}
\description{
collate two vectors together
}
\examples{
intercalate(c("a","d","f"), c("b", "e", "g", "z"))
}
