% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{intercalate_lst}
\alias{intercalate_lst}
\title{collate lists together}
\usage{
intercalate_lst(...)
}
\arguments{
\item{...}{A set of lists to collate}
}
\value{
a lists with elements collated together
}
\description{
interleave multiple vectors
}
\examples{
intercalate(list(paste0(1:5, "_A")), list(paste0(1:3, "_B")), list(paste0(1:6, "_C")))
}
