% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ix_sdiff}
\alias{ix_sdiff}
\title{subset out indices}
\usage{
ix_sdiff(obj, filt_out)
}
\value{
obj with indices indicated in filt_out taken out
}
\description{
A function that subsets out indices and is robust to
if filt_out indices are integer(0)
}
