% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{kpdf}
\alias{kpdf}
\title{open pdf device with defaults}
\usage{
kpdf(
  filename = "plot.pdf",
  height = 10,
  width = 10,
  h = height,
  w = width,
  cex = 1,
  title = NULL,
  byrow = TRUE,
  dim = NULL,
  cex.title = 1,
  oma.scale = 0,
  oma.val = c(1, 1, 1, 1),
  useDingbats = FALSE,
  res = 0,
  pars = list(),
  ...
)
}
\description{
output to pdf device in default directory and keep open
}
