% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ksvg}
\alias{ksvg}
\title{open svg device with ppdf defaults without closing}
\usage{
ksvg(
  filename = "plot.svg",
  height = 10,
  width = 10,
  h = height,
  w = width,
  cex = 1,
  title = NULL,
  byrow = TRUE,
  dim = NULL,
  cex.title = 1,
  oma.scale = 0,
  units = "in",
  oma.val = c(1, 1, 1, 1),
  useDingbats = FALSE,
  res = 300,
  pars = list(),
  ...
)
}
\description{
open svg device with ppdf defaults without closing
}
