% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lapply_dt}
\alias{lapply_dt}
\title{Flexibly apply function to columns of data.table/frame}
\usage{
lapply_dt(x, dt, LFUN = "identity", natype = NA, as.data.table = T)
}
\arguments{
\item{x}{fields to apply function to}

\item{dt}{data.table/frame}

\item{LFUN}{either a character name of a function, or a function}

\item{natype}{the type of NA which is to be specified as one of NA, NA_character_, NA_real_, or NA_integer_}

\item{as.data.table}{a logical indicating whether to coerce the output to a data.table}
}
\value{
data.table/frame if as.data.table is TRUE, otherwise a list
}
\description{
Convenience function to apply a function to columns of a data.table/frame.
The syntax for the columns is flexible

lapply_dt(c(newcolname = colname), dt, dosomething)
will give a data.table or list with the column name
renamed to newcolname
lapply_dt(.(newcolname = colname)...) also works as well.
Note that no quotations are needed.
}
