% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lens}
\alias{lens}
\title{similar to lengths except gets nrows for those items that have dimensions}
\usage{
lens(x, use.names = TRUE)
}
\arguments{
\item{x}{A list}
}
\value{
A numeric vector of lengths of each list
}
\description{
figure out length or nrows of a list
if there are dimensions in the list element,
find out the number of rows
}
