% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{levelsinuse}
\alias{levelsinuse}
\title{get the levels in use in a factor}
\usage{
levelsinuse(fct)
}
\value{
the levels of a factor that are represented in the factor
}
\description{
get the levels in use in a factor
}
