% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{loop_grep}
\alias{loop_grep}
\title{wraps grep in for loop}
\usage{
loop_grep(pattern, x, ignore.case = FALSE)
}
\value{
integer vector of indices
}
\description{
A wrapper around grep to identify string matches across multiple patterns
}
