% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{loop_grepl}
\alias{loop_grepl}
\title{wrapper around loop_grep}
\usage{
loop_grepl(patterns, vec_char, ignore.case = FALSE)
}
\value{
logical vector of all matches
}
\description{
Generating a logical vector from loop_grep
}
