% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lst.empty}
\alias{lst.empty}
\title{lst.empty

A logical vector to select which list elements are empty}
\usage{
lst.empty(x)
}
\value{
A list
}
\description{
lst.empty

A logical vector to select which list elements are empty
}
