% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lst.empty2na}
\alias{lst.empty2na}
\title{empty list elements set to NA

set empty list elements to NA}
\usage{
lst.empty2na(x)
}
\value{
A list
}
\description{
empty list elements set to NA

set empty list elements to NA
}
