% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lst.empty2replace}
\alias{lst.empty2replace}
\alias{lst.emptyreplace}
\title{length 0 elements of list to replace with a value

set empty elements to a replacement value}
\usage{
lst.emptyreplace(x, replace = NA)
}
\value{
A list
}
\description{
length 0 elements of list to replace with a value

set empty elements to a replacement value
}
