% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lst.null2na}
\alias{lst.null2na}
\title{NULL elements of list to NA

set NULL list elements to NA}
\usage{
lst.null2na(x)
}
\value{
A list
}
\description{
NULL elements of list to NA

set NULL list elements to NA
}
