% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{lst.zerochar2empty}
\alias{lst.zerochar2empty}
\title{zero length character elements to empty char

set 0 length chracter to empty}
\usage{
lst.zerochar2empty(x)
}
\value{
A list
}
\description{
zero length character elements to empty char

set 0 length chracter to empty
}
