% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{main}
\alias{main}
\title{wrapper around cenv}
\usage{
main(expr_6000525395_6907698684, return = F)
}
\arguments{
\item{return}{logical}

\item{expr}{expression to evaluate}

\item{env}{environment}
}
\description{
to be able to main
}
