% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{make_chunks}
\alias{make_chunks}
\title{make_chunks}
\usage{
make_chunks(
  vec,
  n = 100,
  max_per_chunk = TRUE,
  num_chunk = !max_per_chunk,
  seed = 10
)
}
\value{
A list
}
\description{
Create chunks from a vector with a certain number of elements per chunk
OR create a certain number of chunks from a vector
}
