% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{make_dummy}
\alias{make_dummy}
\title{make table of dummy encodings}
\usage{
make_dummy(
  x,
  field = ".",
  sep = ".",
  levelsOnly = FALSE,
  fullRank = FALSE,
  return.data.table = T
)
}
\description{

}
