% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{make_ttsplit}
\alias{make_ttsplit}
\title{make training/test splits}
\usage{
make_ttsplit(
  dat,
  field = NULL,
  use.index = TRUE,
  k = 10,
  nested = FALSE,
  times = 1,
  transpose = TRUE,
  seed = 10,
  partition_prob = 0.632,
  split_type = c("crossfold", "resample", "partition"),
  as.data.table = FALSE,
  rngkind = "L'Ecuyer-CMRG"
)
}
\arguments{
\item{dat}{data.table or data.frame of one row per observation. assumes each observation is a row}

\item{k}{number of groups}
}
\value{
a list of row ids corresponding to each fold and training and test split
}
\description{
training and testing
}
\author{
Kevin Hadi
}
