% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{make_xfold}
\alias{make_xfold}
\title{make training/test splits}
\usage{
make_xfold(dat, k = 10, nested = FALSE, times = 1, transpose = TRUE, seed = 10)
}
\arguments{
\item{dat}{data.table or data.frame of one row per observation}

\item{k}{number of groups}
}
\value{
a list of row ids corresponding to each fold and training and test split
}
\description{
make training/test splits
}
\author{
Kevin Hadi
}
