% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{match}
\alias{match}
\title{similar to setkey except a general use utility}
\usage{
match(
  x,
  table,
  nomatch = NA_integer_,
  old = FALSE,
  use.data.table = TRUE,
  return_match = TRUE
)
}
\description{
very slow version of keying a la data.table
but for general/interactive use
}
