% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{match2}
\alias{match2}
\title{matches x in terms of y

returns vector of indices of matches in x with length of vector = length(y)
non matches are NA}
\usage{
match2(x, y)
}
\description{
matches x in terms of y

returns vector of indices of matches in x with length of vector = length(y)
non matches are NA
}
