% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{matrify}
\alias{matrify}
\title{convenience function to convert to matrix
and optionally filter out the first column
which may be rownames that are not relevant to further data analysis}
\usage{
matrify(obj, rm_col1 = TRUE, use.c1.rownames = TRUE)
}
\arguments{
\item{obj}{a data.frame or matrix}

\item{rm_col1}{a logical vector specifying if the 1st column should be removed}
}
\value{
a matrix
}
\description{
convenience function to convert to matrix
and optionally filter out the first column
which may be rownames that are not relevant to further data analysis
}
