% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{matrify2}
\alias{matrify2}
\title{create matrix from table like obj with dplyr syntax}
\usage{
matrify2(df, ..., rownames = 1, rnsep = " ")
}
\description{
create matrix from table like obj with dplyr syntax
}
