% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{max.col.narm}
\alias{max.col.narm}
\title{maximum column per row (removing NA)}
\usage{
\method{max}{col.narm}(mat, ties.method = "first")
}
\description{
Return the index of the maximum column per row while removing NA
}
\author{
Stackoverflow
}
