% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{min.col.narm}
\alias{min.col.narm}
\title{minimum column per row (removing NA)

Return the index of the minimum column per row while removing NA}
\usage{
\method{min}{col.narm}(mat, ties.method = "first")
}
\description{
minimum column per row (removing NA)

Return the index of the minimum column per row while removing NA
}
\author{
stackoverflow
}
