% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mkst}
\alias{mkst}
\title{MaKe STring

making string out of vector for eval(parse(text = ...))}
\usage{
mkst(v, f = "c", po = "(", pc = ")", collapse = ",")
}
\description{
MaKe STring

making string out of vector for eval(parse(text = ...))
}
