% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mrocdat}
\alias{mrocdat}
\title{create a data frame for ggplotting multiroc}
\usage{
mrocdat(lbl, prd)
}
\arguments{
\item{lbl}{output from mroclab}

\item{prd}{output from mrocpred}
}
\value{
A data.frame that can be fed into ggplot
}
\description{
create a data frame for ggplotting multiroc
}
