% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mroclab}
\alias{mroclab}
\title{create one hot table of labels for multiROC}
\usage{
mroclab(y)
}
\arguments{
\item{y}{factor}
}
\value{
the input list
}
\description{
create one hot table of labels for multiROC
}
