% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mrocpred}
\alias{mrocpred}
\title{format predicted scores for multiROC

from predict(...)}
\usage{
mrocpred(prd0, nm = "agg")
}
\arguments{
\item{prd0}{a matrix of prediction scores}
}
\value{
A prediction
}
\description{
format predicted scores for multiROC

from predict(...)
}
