% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg}
\alias{msg}
\title{msg}
\usage{
msg(..., .silent = getOption("khtools.silent", FALSE))
}
\arguments{
\item{...}{objekter som skal skrives ut}

\item{.silent}{mulighet til å skru av logging ved testing}
}
\description{
Skriver melding til konsoll og flusher output.
}
\concept{utilities}
