% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mstrsplit}
\alias{mstrsplit}
\title{make matrix out of stringsplitted character vector

split a vector by a delimiter,
fill in to make same length
bind into matrix}
\usage{
mstrsplit(x, ...)
}
\value{
a matrix
}
\description{
make matrix out of stringsplitted character vector

split a vector by a delimiter,
fill in to make same length
bind into matrix
}
