% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{na2empty}
\alias{na2empty}
\title{na2empty

A convenience function to set a character vector with NAs to an
empty character}
\usage{
na2empty(v)
}
\value{
A character vector
}
\description{
na2empty

A convenience function to set a character vector with NAs to an
empty character
}
