% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{na2true}
\alias{na2true}
\title{replace logical vector with NA to TRUE}
\usage{
na2true(v)
}
\value{
A logical vector with NAs set to TRUE
}
\description{
A convenience function to set a logical vector with NAs to TRUE
}
