% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{na2zero}
\alias{na2zero}
\title{na2zero

A convenience function to set a numeric vector with NAs to zero}
\usage{
na2zero(v)
}
\value{
A numeric vector with NAs set to zero
}
\description{
na2zero

A convenience function to set a numeric vector with NAs to zero
}
