% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{names2<-}
\alias{names2<-}
\title{robust name() assignment}
\usage{
names2(x, useempty = FALSE) <- value
}
\arguments{
\item{x}{vector}
}
\value{
x a vector with all names
}
\description{
similar to rlang::names2
}
