% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{nan2zero}
\alias{nan2zero}
\title{nan2zero

A convenience function to set a numeric vector with NaNs to zero}
\usage{
nan2zero(v)
}
\value{
A numeric vector with NaNs set to zero
}
\description{
nan2zero

A convenience function to set a numeric vector with NaNs to zero
}
