% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{nonzero2na}
\alias{nonzero2na}
\title{replace 0 to with NA}
\usage{
nonzero2na(x)
}
\value{
A logical vector with NAs set to FALSE
}
\description{
A convenience function to set a logical vector with NAs to false
}
