% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{normv}
\alias{normv}
\title{normalize a vector

i.e. rescale to have values between 0 and 1}
\usage{
normv(x)
}
\value{
vector
}
\description{
normalize a vector

i.e. rescale to have values between 0 and 1
}
