% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{normv_sep}
\alias{normv_sep}
\title{normalize a vector, treating positives and negatives separately

i.e. rescale negatives to be between 0-0.5
rescale positives to be between 0.5-1}
\usage{
normv_sep(x)
}
\value{
vector
}
\description{
normalize a vector, treating positives and negatives separately

i.e. rescale negatives to be between 0-0.5
rescale positives to be between 0.5-1
}
