% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{numeq}
\alias{numeq}
\title{test equality between numeric values with some tolerance}
\usage{
numeq(x, y, tol = .Machine$double.eps^0.5)
}
\value{
logical vector
}
\description{
two numerical values may be slightly off in their decimal precision.
These may be considered equivalent values but the `==` operator will
return FALSE. This tests for equivalence of two values with some lower
tolerance limit
}
