% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{output_cols}
\alias{output_cols}
\title{output the union columns from a Flow output}
\usage{
output_cols(x, mc.cores = 1)
}
\value{
character vector of all output columns to be expected
}
\description{
output the union columns from a Flow output
}
