% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{overwritefun}
\alias{overwritefun}
\title{overwrite a function in its namespace

useful for dev purposes.}
\usage{
overwritefun(newfun, oldfun, package, envir = globalenv())
}
\description{
overwrite a function in its namespace

useful for dev purposes.
}
