% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{overwriteR6}
\alias{overwriteR6}
\title{overwrite a method in R6 class generator}
\usage{
overwriteR6(
  newfun,
  oldfun,
  r6gen,
  meth = "public_methods",
  package = NULL,
  envir = globalenv()
)
}
\description{
useful for dev purposes.
}
