% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pairs.collect.junctions}
\alias{pairs.collect.junctions}
\title{collect junctions from pairs as breakpoints}
\usage{
\method{pairs}{collect.junctions}(
  pairs,
  jn.field = "complex",
  id.field = "pair",
  mc.cores = 1,
  mask = "/gpfs/commons/groups/imielinski_lab/DB/Broad/um75-hs37d5.bed.gz",
  ev.types = c("bfb", "chromoplexy", "chromothripsis", "del", "dm", "dup", "fbi",
    "pyrgo", "qrdup", "qrdel", "qrp", "rigma", "simple_inv", "simple_invdup",
    "simple_tra", "tic", "tyfonas", "cpxdm", "tib")
)
}
\description{

}
