% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pairs.grab.hrdetect.features}
\alias{pairs.grab.hrdetect.features}
\title{grab hrdetect features from pairs table}
\usage{
\method{pairs}{grab.hrdetect.features}(pairs, field = "hrd_results", id.field = "pair", mc.cores = 1)
}
\description{
wrapper around grab.hrdetect.features
}
