% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pairs.grab.ot.features}
\alias{pairs.grab.ot.features}
\title{grab oneness twoness features from pairs table}
\usage{
\method{pairs}{grab.ot.features}(pairs, field = "oneness_twoness", id.field = "pair", mc.cores = 1)
}
\description{
wrapper around grab.ot.features
}
