% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{parasn}
\alias{parasn}
\title{assign parallel columns}
\usage{
parasn(x, y, cols, sans_key = TRUE, use.data.table = T)
}
\arguments{
\item{x}{data.table}

\item{y}{data.table}
}
\description{
assign parallel columns
}
\author{
Kevin Hadi
}
