% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{parse.grl2}
\alias{parse.grl2}
\title{a robust parse.grl}
\usage{
parse.grl2(str, meta = NULL, fixna = FALSE)
}
\arguments{
\item{str}{A string that can be parsed as ranged data "A:1-10+"}

\item{meta}{A table that will be added as metadata}
}
\value{
GRangesList
}
\description{
version of parse.grl that is able to convert ranges with minus signs
}
\author{
Kevin Hadi
}
