% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{parsesnpeff}
\alias{parsesnpeff}
\title{parse snpeff output into granges}
\usage{
parsesnpeff(
  vcf,
  id = NULL,
  filterpass = TRUE,
  coding_alt_only = TRUE,
  geno = NULL,
  gr = NULL,
  keepfile = FALSE,
  altpipe = FALSE,
  debug = FALSE,
  snpeffpath = "~/modules/SnpEff",
  filters = "PASS,."
)
}
\arguments{
\item{vcf}{path to snpeff vcf}

\item{pad}{Exposed argument to skitools::ra.overlaps()}
}
\value{
GRangesList of breakpoint pairs with junctions that overlap removed
}
\description{
parse snpeff output into granges
}
