% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pinch.frac}
\alias{pinch.frac}
\title{pinch a vector

A convenience function to transform proportions.
Useful for beta regression (library(betareg))}
\usage{
pinch.frac(x, fmin = 0.01, fmax = 0.99)
}
\value{
A vector
}
\description{
pinch a vector

A convenience function to transform proportions.
Useful for beta regression (library(betareg))
}
