% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ppdf}
\alias{ppdf}
\title{kevin's modification of ppdf

height and width are specified in inches by default
resolution is specified as 300 but is not used}
\usage{
ppdf(
  expr,
  filename = "plot.pdf",
  height = 10,
  width = 10,
  h = height,
  w = width,
  cex = 1,
  title = NULL,
  byrow = TRUE,
  dim = NULL,
  cex.title = 1,
  oma.scale = 0,
  oma.val = c(1, 1, 1, 1),
  useDingbats = FALSE,
  res = 0,
  ...
)
}
\description{
kevin's modification of ppdf

height and width are specified in inches by default
resolution is specified as 300 but is not used
}
