% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{process_alt_for_hrd}
\alias{process_alt_for_hrd}
\title{process mutlist from process_alterations}
\usage{
process_alt_for_hrd(
  mutlist,
  gpath = "~/lab/projects/Starr/BRCA/files/hrd_gene_list_jsetton.csv"
)
}
\description{
process mutlist from process_alterations
}
