% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{process_tbl}
\alias{process_tbl}
\title{flexibly read in a field and append an id to the output}
\usage{
process_tbl(
  tbl,
  field = "jabba_rds",
  id.field = "pair",
  read.fun,
  remove_ext = c(".gz", ".zip"),
  mc.cores = 1
)
}
\arguments{
\item{tbl}{table with fields to read in}

\item{field}{field to read in}

\item{id.field}{field with id to append to output}

\item{read.fun}{function to read in, will try to guess based on extension, but may need to provide}

\item{remove_ext}{extension strings to remove from file}

\item{mc.cores}{number of cores}
}
\value{
a list of idx and seq
}
\description{
perform an embarrassingly parallel operation on files
i.e. read in and do some pre processing
}
