% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{qtrim}
\alias{qtrim}
\title{cut off vector by quantiles}
\usage{
qtrim(x, maxq = 0.99)
}
\description{
Use the empirical distribution of numeric values
to get the quantiles assigned to each value of a vector.
then remove values that are above the quantile cutoff.
}
