% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{read.bam.header}
\alias{read.bam.header}
\title{read.bam.header

Read in a bam header into tabular format}
\usage{
read.bam.header(bam, trim = FALSE)
}
\value{
A data.table
}
\description{
read.bam.header

Read in a bam header into tabular format
}
