% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{read_bed}
\alias{read_bed}
\title{read bed or bedpe as a table

Read in a bed oe bedpe formatted file into tabular format}
\usage{
read_bed(bedpath)
}
\value{
A data.table
}
\description{
read bed or bedpe as a table

Read in a bed oe bedpe formatted file into tabular format
}
