% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{read_vcf2}
\alias{read_vcf2}
\title{read_vcf2

read in a vcf file to granges with additional processing with bcftools}
\usage{
read_vcf2(
  fn,
  gr = NULL,
  type = c("snps", "indels", "all"),
  hg = "hg19",
  geno = NULL,
  swap.header = NULL,
  verbose = FALSE,
  add.path = FALSE,
  tmp.dir = tempdir(),
  ...
)
}
\value{
GRanges
}
\description{
read_vcf2

read in a vcf file to granges with additional processing with bcftools
}
