% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{readin}
\alias{readin}
\title{flexible file opening}
\usage{
readin(
  x,
  txt.fun = data.table::fread,
  vcf.fun = skidb::read_vcf,
  other.txt = NULL,
  alt.fun = NULL,
  alt.ext = NULL,
  other.compress = NULL
)
}
\description{
open text, vcf, or rds files flexibly
}
