% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{readinfasta}
\alias{readinfasta}
\title{wrapper around readDNAStringSet to remove extraneous characters}
\usage{
readinfasta(fa, allow_vertbar = FALSE)
}
\value{
DNAStringSet object
}
\description{
Uses Biostrings::readDNAStringSet to read in fasta
and remove all characters that do not correspond to rname
}
\author{
Kevin Hadi
}
