% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{reassign}
\alias{reassign}
\title{reassign elements of named list into environment}
\usage{
reassign(variables_lst, calling_env = parent.frame())
}
\arguments{
\item{variables_lst}{A named list of variables}
}
\value{
the input list
}
\description{
ONLY USE IF YOU KNOW WHAT YOU ARE DOING
this function takes a list of named objects and assigns them to the calling environment
}
