% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{refactor}
\alias{refactor}
\title{refactor

Keep one level of a factor and set all others to a specified level}
\usage{
refactor(fac, keep, ref_level = "OTHER")
}
\value{
A factor
}
\description{
refactor

Keep one level of a factor and set all others to a specified level
}
