% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rep_each}
\alias{rep_each}
\title{recycle vector - shortcut for rep(x, each = each)}
\usage{
rep_each(x, each)
}
\arguments{
\item{x}{data}

\item{each}{length to extend vector by}
}
\value{
vector
}
\description{

}
\author{
Kevin Hadi
}
