% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rep_len}
\alias{rep_len}
\title{recycle vector - overload base::rep_len}
\usage{
rep_len(x, length.out)
}
\arguments{
\item{x}{data}

\item{length.out}{length to extend vector by}
}
\value{
vector
}
\description{
problem with base::rep_len is that it doesn't work with other objects
}
\author{
Kevin Hadi
}
