% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{replace2}
\alias{replace2}
\title{function to replace elements of vector, can use "x" as surrogate variable in expression}
\usage{
replace2(x, repl.expr, values)
}
\description{
convenience function to replace without having to type excessively
if the variable is arrived at through nested functions or long
variable names
}
