% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{reset.job}
\alias{reset.job}
\title{reset.job}
\usage{
reset.job(
  x,
  ...,
  i = NULL,
  rootdir = x@rootdir,
  jb.mem = x@runinfo$mem,
  jb.cores = x@runinfo$cores,
  jb.time = x@runinfo$time,
  update_cores = 1,
  task = NULL
)
}
\value{
A Flow job object
}
\description{
Reset a job with different params
}
