% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ret_no_err}
\alias{ret_no_err}
\title{a wrapper around check_lst}
\usage{
ret_no_err(lst, class_condition = c("try-error", "error", "errored",
  "err"))
}
\arguments{
\item{lst}{A list (usually the output of lapply(... , function(x) try({}))}
}
\value{
only returns the non-errors in the list
}
\description{
a wrapper around check_lst
}
