% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rg_sub}
\alias{rg_sub}
\title{extracting substring match using regexpr}
\usage{
rg_sub(pattern, text, ...)
}
\value{
Character vector of the regex matched portions of the input string vector
}
\description{
extract the first portion of matched substring
}
