% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rmcol}
\alias{rmcol}
\title{rmcol}
\usage{
rmcol(df, rmcol = 1, usekey = T, invert = T, exact = FALSE)
}
\arguments{
\item{df}{object with !is.null(df) == TRUE}
}
\description{
rmcol
}
\author{
Kevin Hadi
}
