% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_utils.R
\name{robust.scale}
\alias{robust.scale}
\title{flatten outliers by quantile}
\usage{
robust.scale(x, qlow = 0.1, qhigh = 0.9)
}
\value{
a vector of features pmin and pmax by quantile
}
\description{
enforce outliers to pmax and pmin within a quantile
}
