% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rownames2<-}
\alias{rownames2<-}
\title{robust rownames() assignment}
\usage{
rownames2(x, useempty = FALSE) <- value
}
\arguments{
\item{x}{vector or matrix}
}
\value{
x a vector fully named or rownamed
}
\description{
robust rownames() assignment
}
