% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rownames_to_column}
\alias{rownames_to_column}
\title{making column out of rownames

internal version that doesn't require library(tibble)}
\usage{
rownames_to_column(.data, var = "rowname")
}
\arguments{
\item{.data}{a data frame/table}
}
\value{
a data frame/table with the rownames as an additional column
}
\description{
making column out of rownames

internal version that doesn't require library(tibble)
}
