% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{sampler}
\alias{sampler}
\title{sampler}
\usage{
sampler(
  x,
  n = NROW(x),
  seed = 10,
  rngkind = "L'Ecuyer-CMRG",
  verbose = FALSE,
  replace = FALSE,
  prob = NULL
)
}
\description{
sample elements of vector or rows of table
}
