% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{select.matrix}
\alias{select.matrix}
\title{select.matrix}
\usage{
select.matrix(x, rows = NULL, cols = NULL, int.rows = TRUE, int.cols = TRUE)
}
\value{
matrix
}
\description{
wrapper to pick out rows and columns without erroring out
}
