% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{selfname}
\alias{selfname}
\title{name a character vector to itself}
\usage{
selfname(char)
}
\arguments{
\item{char}{A character vector}
}
\value{
A named character vector
}
\description{
self explanatory
}
