% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{set_rngseed}
\alias{set_rngseed}
\title{set random number generator AND seed}
\usage{
set_rngseed(
  seed = 10,
  rngkind = "L'Ecuyer-CMRG",
  normal.kind = "Inversion",
  sample.kind = "Rejection",
  use.old.sample.kind = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{seed}{integer specifying seed}

\item{rngkind}{string specifying number generator}

\item{use.old.sample.kind}{logical specifying discrete uniform generation method from prior to R361}
}
\value{
a list of row ids corresponding to each fold and training and test split
}
\description{
set random number generator AND seed
}
\author{
Kevin Hadi
}
