% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{setNames2}
\alias{setNames2}
\title{setNames2

convenience function to set all names of a vector}
\usage{
setNames2(vec, nm, useempty = FALSE)
}
\arguments{
\item{vec}{vector}

\item{nm}{character}
}
\value{
named vector
}
\description{
setNames2

convenience function to set all names of a vector
}
