% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{setRownames}
\alias{setRownames}
\title{convenience function to set row names}
\usage{
setRownames(object = nm, nm)
}
\arguments{
\item{object}{tabled object}

\item{nm}{names of the new columns}
}
\value{
rownamed object
}
\description{
convenience function to set row names
}
