% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{shift_down}
\alias{shift_down}
\title{shift_downstream from plyranges package}
\usage{
shift_down(x, shift = 0L)
}
\value{
granges
}
\description{
works on genomicranges
}
\author{
Stuart Lee

Michael Lawrence

Dianne Cook
}
