% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{shift_left}
\alias{shift_left}
\title{shift_left from plyranges package}
\usage{
shift_left(x, shift = 0L)
}
\value{
granges
}
\description{
works on genomicranges
}
\author{
Stuart Lee

Michael Lawrence

Dianne Cook
}
