% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{shift_right}
\alias{shift_right}
\title{shift_right from plyranges package}
\usage{
shift_right(x, shift = 0L)
}
\value{
granges
}
\description{
works on genomicranges
}
\author{
Stuart Lee

Michael Lawrence

Dianne Cook
}
