% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{sim.bx}
\alias{sim.bx}
\title{simulate 10x Barcoded reads off of collection of walks}
\usage{
sim.bx(
  wk,
  only_simulate_breaks = TRUE,
  numr = 24,
  bxwid = 30000,
  physcov = 150,
  readlength = 150,
  fraglength = 1200
)
}
\description{
Take a collection of walks and simulate barcoded reads at the break sites
}
\author{
Kevin Hadi
}
