% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sql_quote_I}
\alias{sql_quote_I}
\title{sql_quote_I}
\usage{
sql_quote_I(con, x)
}
\arguments{
\item{con}{db connection}

\item{x}{object to quote as identifier}
}
\description{
DBI::dbQuoteIdentifier
}
\seealso{
Other sql:
\code{\link[=sql_quote_S]{sql_quote_S()}}
}
\concept{sql}
