% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sql_quote_S}
\alias{sql_quote_S}
\title{sql_quote_S}
\usage{
sql_quote_S(con, x)
}
\arguments{
\item{con}{db connection}

\item{x}{object to quote as string}
}
\description{
DBI::dbQuoteString
}
\seealso{
Other sql:
\code{\link[=sql_quote_I]{sql_quote_I()}}
}
\concept{sql}
