% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ss}
\alias{ss}
\title{same as subset2

convenience function to subset without having to type excessively
if the variable is arrived at through nested functions or long
variable names}
\usage{
ss(x, sub.expr, ...)
}
\description{
same as subset2

convenience function to subset without having to type excessively
if the variable is arrived at through nested functions or long
variable names
}
\author{
Kevin Hadi
}
