% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{stack.dt}
\alias{stack.dt}
\title{collapse a named list of vectors into a data.table

Collapse a named list with vectors as each element into a data.table}
\usage{
stack.dt(lst, ind = "ind", values = "values", ind.as.character = TRUE)
}
\description{
collapse a named list of vectors into a data.table

Collapse a named list with vectors as each element into a data.table
}
