% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{sv_filter}
\alias{sv_filter}
\title{filter sv by overlaps with another}
\usage{
sv_filter(sv, filt_sv, pad = 500)
}
\arguments{
\item{sv}{GRangesList with all elements length 2 (specifying breakpoint pairs of a junction)}

\item{filt_sv}{GRangesList with all elements length 2 (usually a pon)}

\item{pad}{Exposed argument to skitools::ra.overlaps()}
}
\value{
GRangesList of breakpoint pairs with junctions that overlap removed
}
\description{
To filter out a grangeslist of sv by another grangeslist of SV
}
