% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{symdiff}
\alias{symdiff}
\title{data.table of all setdiff items in X and in Y

gives back data table of setdiff elements
noting whether the element is in vector x or vector y
This gives back all elements, including non-unique}
\usage{
symdiff(x, y, ignore.na = FALSE)
}
\value{
data.table
}
\description{
data.table of all setdiff items in X and in Y

gives back data table of setdiff elements
noting whether the element is in vector x or vector y
This gives back all elements, including non-unique
}
