% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{system3}
\alias{system3}
\title{modified system2

A modification of system2 to be able to return either or the stderr
or stdout to console or flexibly return to a path.}
\usage{
system3(
  command,
  args = character(),
  stdout = "",
  stderr = "",
  stdin = "",
  input = NULL,
  env = character(),
  wait = TRUE,
  minimized = FALSE,
  invisible = TRUE,
  timeout = 0
)
}
\description{
modified system2

A modification of system2 to be able to return either or the stderr
or stdout to console or flexibly return to a path.
}
