% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{trans.df}
\alias{trans.df}
\title{transpose data.table or data.frame}
\usage{
trans.df(df, rn = 1)
}
\description{
transpose a data frame
by default, the first column is stripped
and used as colnames of the output table
}
