% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{try2}
\alias{try2}
\title{wrapper around tryCatch - robust to parallel:: functions}
\usage{
try2(expr, ..., finally)
}
\description{
A slightly more robust version of try that works within the parallel:: set of functions
that pre-deploy a cluster.
}
