% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{undebug.s4}
\alias{undebug.s4}
\title{undebug an S4 function

wrapper around untrace}
\usage{
undebug.s4(what, signature, where)

undebug.s4(what, signature, where)
}
\description{
undebugging an S4 function when debug.s4/trace were
called on an S4 method.
}
