% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{undup}
\alias{undup}
\title{an alternative to base::unique() that preserves names}
\usage{
undup(obj, fromLast = FALSE, nmax = NA, na.rm = FALSE)
}
\arguments{
\item{obj}{an R vector}
}
\value{
unique values of obj with names preserved
}
\description{
an alternative to base::unique() that preserves names
}
