% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{upset2}
\alias{upset2}
\title{wrapper around UpSetR::upset()}
\usage{
upset2(
  data,
  text.scale = 1.5,
  mb.ratio = c(0.7, 0.3),
  empty.intersections = "on",
  ...
)
}
\description{
convenience function for better UpSetR defaults
}
