% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{vcf_remove_sample}
\alias{vcf_remove_sample}
\title{take vcf read in chunks and write clean vcf

take vcf read in chunks and write clean vcf without sample info
i.e. fixing up svaba indel crap}
\usage{
vcf_remove_sample(
 
    x = "/gpfs/commons/groups/imielinski_lab/data/PCAWG/mutations/f393baf9-2710-9203-e040-11ac0d484504,vcf",
  out.vcf = "~/outtest.vcf",
  ref = "~/DB/GATK/hg19_gatk_decoy.fasta",
  chunk = 10000,
  verbose = TRUE
)
}
\value{
character
}
\description{
take vcf read in chunks and write clean vcf

take vcf read in chunks and write clean vcf without sample info
i.e. fixing up svaba indel crap
}
