% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{viewtask}
\alias{viewtask}
\title{convert job task to table}
\usage{
viewtask(jb, arglst = c("name", "arg", "default"))
}
\value{
A Flow job object
}
\description{
Parse task from Flow Job object, character task file,
or Flow Task object
}
