% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{vmatch}
\alias{vmatch}
\title{make list of matches and nonmatches

get match indices and non matched indices}
\usage{
vmatch(x, y, ...)
}
\description{
make list of matches and nonmatches

get match indices and non matched indices
}
