% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{whicherr}
\alias{whicherr}
\title{same as iderr}
\usage{
whicherr(
  lst,
  class_condition = c("simpleError", "try-error", "error", "errored", "err")
)
}
\arguments{
\item{lst}{A list}
}
\value{
a logical vector marking which elements are try-errors"
}
\description{
checking a list for any elements that are try-errors
usually from an lapply(..., function(x) try({})) call
}
