% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{withv}
\alias{withv}
\title{withv

to be used for quick interactive programming
withv(toolongtotypemeagain, x * sum(x))}
\usage{
withv(x, expr)
}
\description{
withv

to be used for quick interactive programming
withv(toolongtotypemeagain, x * sum(x))
}
