% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{withx}
\alias{withx}
\title{withx

to be used for quick interactive programming
withx(toolongtotypemeagain, x * sum(x))}
\usage{
withx(x, expr)
}
\description{
withx

to be used for quick interactive programming
withx(toolongtotypemeagain, x * sum(x))
}
