% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{write.ctab}
\alias{write.ctab}
\title{writing a comma separated table with quotes

comma-separated table with quotes around strings}
\usage{
write.ctab(x, ..., sep = ",", quote = T, row.names = F)
}
\description{
writing a comma separated table with quotes

comma-separated table with quotes around strings
}
