% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ws2und}
\alias{ws2und}
\title{Clean up whitespace from columns of data.frame-like object}
\usage{
ws2und(df)
}
\value{
A character vector
}
\description{
make data.frame or data.table column name whitespaces into
underscores and remove end whitespaces
}
