% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{wv}
\alias{wv}
\title{alias for withv

to be used for quick interactive programming
withv(toolongtotypemeagain, x * sum(x))}
\usage{
wv(x, expr)
}
\description{
alias for withv

to be used for quick interactive programming
withv(toolongtotypemeagain, x * sum(x))
}
