% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{clobber}
\alias{clobber}
\title{same as dplyr::coalesce

clobber NA, or some value between multiple vectors
bads can be a function that returns a logical}
\usage{
clobber(
  ...,
  bads = NA,
  bads.x = NULL,
  bads.y = NULL,
  r2l = FALSE,
  fromLast = FALSE,
  opposite = TRUE
)
}
\arguments{
\item{...}{vectors to merge together}

\item{bads}{a set of values to clobber, or a function that returns a logical}

\item{r2l}{merge from left to right per pair of vectors}

\item{fromLast}{if TRUE, merge from last vector to first}
}
\description{
same as dplyr::coalesce

clobber NA, or some value between multiple vectors
bads can be a function that returns a logical
}
