% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{col2rn}
\alias{col2rn}
\title{alias for column_to_rownames

internal version that doesn't require library(tibble)}
\usage{
col2rn(.data, var = "rowname")
}
\arguments{
\item{.data}{a data frame/table}
}
\value{
a data frame/table with rownames from a column
}
\description{
alias for column_to_rownames

internal version that doesn't require library(tibble)
}
