% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{column_to_rownames}
\alias{column_to_rownames}
\title{making column into rownames

internal version that doesn't require library(tibble)}
\usage{
column_to_rownames(.data, var = "rowname", force = T, sep = " ")
}
\arguments{
\item{.data}{a data frame/table}
}
\value{
a data frame/table with rownames from a column
}
\description{
making column into rownames

internal version that doesn't require library(tibble)
}
