% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dcast.count}
\alias{dcast.count}
\title{dcast.count}
\usage{
dcast.count(tbl, lh, rh = NULL, countcol = "count", ...)
}
\value{
A data frame or data.table
}
\description{
Counting up occurrences in a table while taking factor levels into account
}
