% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{dt_empty2na}
\alias{dt_empty2na}
\alias{dt_setnull}
\title{convert columns with empty character to NA

coerce "" in columns of class "character" to NA_character_}
\usage{
dt_empty2na(dt)

dt_setnull(dt, cols)
}
\arguments{
\item{dt}{data.table}

\item{cols}{character vector of column names}

\item{these_cols}{NULL by default, will select columns of character class, otherwise will be specified}
}
\value{
A data.table

A data.table
}
\description{
convert columns with empty character to NA

coerce "" in columns of class "character" to NA_character_

set columnn to NULL in data.table

remove column from data table by setting to NULL
}
