% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_read.R
\name{duckdb_fetch_table}
\alias{duckdb_fetch_table}
\title{duckdb_fetch_table}
\usage{
duckdb_fetch_table(con, tablename, limit = NULL)
}
\arguments{
\item{con}{db connection}

\item{tablename}{navnet på tabellen du skal sjekke}

\item{limit}{antall rader du vil lese (f.eks. limit = 10 for å bare lese øverste 10 rader)}
}
\value{
data.table object
}
\description{
fetch table from duckdb
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_and_replace_table]{duckdb_create_and_replace_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
