% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_utils.R
\name{duckdb_merge_tables}
\alias{duckdb_merge_tables}
\title{duckdb_merge_tables}
\usage{
duckdb_merge_tables(con, mergeto, mergefrom, join_cols = NULL, result = NULL)
}
\arguments{
\item{con}{db connection}

\item{mergeto}{tabellen som skal motta nye kolonner}

\item{mergefrom}{tabellen det skal merges fra}

\item{join_cols}{kolonner som skal brukes som join-nøkler}

\item{result}{resultattabell. Default er \code{mergeto}}
}
\description{
Merger nye kolonner fra én tabell inn i en annen.

Kolonner som finnes i begge tabeller benyttes som
potensielle join-kolonner. Kun kolonner angitt i
\code{join_cols} og som finnes i begge tabeller brukes
i join-betingelsen.

Dersom \code{result = mergeto} erstattes tabellen ved hjelp
av en midlertidig tabell og \code{duckdb_replace_table()}.
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_and_replace_table]{duckdb_create_and_replace_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
