% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_write.R
\name{duckdb_replace_table}
\alias{duckdb_replace_table}
\title{replace_table_duckdb}
\usage{
duckdb_replace_table(con, target, source)
}
\arguments{
\item{con}{db connection}

\item{target}{navn på tabellen man skal sitte igjen med}

\item{source}{navn på tabellen som skal ende opp som target}
}
\description{
Erstatter en tabell med en annen i duckdb. Ved bearbeiding av en tabell kan resultatet
skrives til en tmp-tabell, og så kan hovedtabellen erstattes med denne etterpå. Da slipper
vi å overskrive tabellen med seg selv som kan være ustabilt.
Vi kan i stedet bruke CREATE TABLE tmp AS SELECT * FROM TABELL, og deretter bruke
replace_table_duckdb(con, target = TABELL, source = tmp). Dette vil først generere ny tabell
tmp, og deretter erstatte originaltabellen med denne.
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_and_replace_table]{duckdb_create_and_replace_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
