% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_save.R
\name{duckdb_save_file}
\alias{duckdb_save_file}
\title{duckdb_save_file}
\usage{
duckdb_save_file(con, source, filepath)
}
\arguments{
\item{con}{db connection}

\item{source}{tabellnavn eller SQL-uttrykk}

\item{filepath}{filsti}
}
\description{
Eksporterer en tabell eller et SQL-uttrykk fra DuckDB til fil.

Wrapper rundt DuckDB COPY ... TO.
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_and_replace_table]{duckdb_create_and_replace_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
