% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_write.R
\name{duckdb_write_and_replace_table_from_R}
\alias{duckdb_write_and_replace_table_from_R}
\title{duckdb_write_and_replace_table}
\usage{
duckdb_write_and_replace_table_from_R(con, tablename, data, temporary = TRUE)
}
\arguments{
\item{con}{db connection}

\item{tablename}{tabellen som skal erstattes}

\item{data}{data.frame eller data.table som skal skrives}

\item{temporary}{om tmp-tabellen skal være temporary}
}
\description{
Skriver data til en midlertidig tabell og erstatter deretter
måltabellen atomisk ved hjelp av duckdb_replace_table().
Dette er nyttig når en tabell er generert i R, eller lest ut av DuckDB, bearbeidet i R,
og skal skrives tilbake uten å overskrive tabellen med seg selv som kan være ustabilt.
}
\seealso{
Other duckdb:
\code{\link[=duckdb_append_table]{duckdb_append_table()}},
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_new_table]{duckdb_create_new_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_column_types]{duckdb_get_column_types()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_existing_table]{duckdb_replace_existing_table()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_table]{duckdb_write_table()}}
}
\concept{duckdb}
