% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_write.R
\name{duckdb_write_table}
\alias{duckdb_write_table}
\title{duckdb_write_table}
\usage{
duckdb_write_table(con, tablename, data, temp = TRUE, overwrite = TRUE, ...)
}
\arguments{
\item{con}{db connection}

\item{tablename}{navn på tabellen du skal skrive}

\item{data}{datasett som skal skrives}

\item{temp}{er det en midlertidig tabell, default = TRUE}

\item{overwrite}{oveskrive dersom tabellen eksisterer, default = TRUE}

\item{...}{andre argumenter som aksepteres av DBI::dbWriteTable, som field.types for å spesifisere kolonnetype}
}
\description{
Skriv eller overskriv eksisterende tabell i duckdb
}
\seealso{
Other duckdb:
\code{\link[=duckdb_clean]{duckdb_clean()}},
\code{\link[=duckdb_create_new_table]{duckdb_create_new_table()}},
\code{\link[=duckdb_drop_tables]{duckdb_drop_tables()}},
\code{\link[=duckdb_ensure_columns]{duckdb_ensure_columns()}},
\code{\link[=duckdb_fetch_table]{duckdb_fetch_table()}},
\code{\link[=duckdb_get_columns]{duckdb_get_columns()}},
\code{\link[=duckdb_get_tables]{duckdb_get_tables()}},
\code{\link[=duckdb_init]{duckdb_init()}},
\code{\link[=duckdb_merge_tables]{duckdb_merge_tables()}},
\code{\link[=duckdb_replace_existing_table]{duckdb_replace_existing_table()}},
\code{\link[=duckdb_replace_table]{duckdb_replace_table()}},
\code{\link[=duckdb_save_file]{duckdb_save_file()}},
\code{\link[=duckdb_shutdown]{duckdb_shutdown()}},
\code{\link[=duckdb_table_exists]{duckdb_table_exists()}},
\code{\link[=duckdb_write_and_replace_table_from_R]{duckdb_write_and_replace_table_from_R()}}
}
\concept{duckdb}
