% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{est_snv_cn_stub}
\alias{est_snv_cn_stub}
\title{estimate snv cn stub}
\usage{
est_snv_cn_stub(
  vcf,
  jab,
  tumbam = NULL,
  germ_subsample = 2e+05,
  somatic = FALSE,
  saveme = FALSE
)
}
\arguments{
\item{vcf}{path to vcf}

\item{jab}{path to jabba}
}
\description{
estimate snv cn stub
}
