% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{file.exists2}
\alias{file.exists2}
\title{slightly more robust test for whether file exists


test whether a file is NOT NA, NULL, or /dev/null and if 
the file exists}
\usage{
file.exists2(x, nullfile = "/dev/null")
}
\arguments{
\item{x}{a character vector}
}
\value{
logical
}
\description{
slightly more robust test for whether file exists


test whether a file is NOT NA, NULL, or /dev/null and if 
the file exists
}
\author{
Kevin Hadi
}
