% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_utils.R
\name{fitglmnet}
\alias{fitglmnet}
\title{fit glmnet}
\usage{
fitglmnet(
  var,
  lambda = 0,
  family = "multinomial",
  ycol,
  dat,
  only_positive = TRUE,
  maxit = 1e+05,
  thresh = 1e-07,
  standardize = FALSE
)
}
\description{
fit glmnet
}
