% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gbar.error}
\alias{gbar.error}
\title{barplot with errorbars

Barplot with confidence intervals.
To plot with confidence intervals around an event that has a binary
outcome with a simple example:

heads = 30, tails = 20,
y = heads / (heads + tails)
dt = binom.conf(y, heads + tails)
conf.low = dt$conf.low
conf.high = dt$conf.high

gbar.error(y, conf.low, conf.high)}
\usage{
gbar.error(
  y,
  conf.low,
  conf.high,
  group,
  wes = "Royal1",
  other.palette = NULL,
  print = TRUE,
  fill = NULL,
  stat = "identity",
  facet1 = NULL,
  facet2 = NULL,
  bar.width = 0.9,
  position = position_dodge(width = 0.9),
  transpose = FALSE,
  facet.scales = "fixed"
)
}
\arguments{
\item{y}{any numeric vector, can be a fraction}

\item{conf.low}{the lower bound of the confidence interval around y}

\item{conf.high}{the upper bound of the confidence interval around y}
}
\value{
A ggplot object
}
\description{
barplot with errorbars

Barplot with confidence intervals.
To plot with confidence intervals around an event that has a binary
outcome with a simple example:

heads = 30, tails = 20,
y = heads / (heads + tails)
dt = binom.conf(y, heads + tails)
conf.low = dt$conf.low
conf.high = dt$conf.high

gbar.error(y, conf.low, conf.high)
}
