% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{getbp}
\alias{getbp}
\title{get breakpoints}
\usage{
getbp(
  gg,
  ignore.small = T,
  ignore.isolated = T,
  max.small = 10000,
  min.isolated = max.small,
  sort = FALSE
)
}
\description{
grab breakpoints from gGraph alt edges and mark with ecluster filters
}
