% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{gr.resize}
\alias{gr.resize}
\title{Resize granges without running into negative width error}
\usage{
gr.resize(gr, wid, each = TRUE, resize = TRUE, fix = "center")
}
\value{
GRanges
}
\description{
lower size limit of window is 0
}
\author{
Kevin Hadi
}
