% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khtools-package.R
\docType{package}
\name{khtools-package}
\alias{khtools}
\alias{khtools-package}
\title{khtools: Verktøy som brukes på tvers i produksjonsapparatet}
\description{
Funksjoner som brukes på tvers av produksjonsapparatet
}
\author{
\strong{Maintainer}: Vegard Lysne \email{vegard.lysne@helsedir.no}

Authors:
\itemize{
  \item Vegard Lysne \email{vegard.lysne@helsedir.no}
}

}
\keyword{internal}
