% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khtools-package.R
\docType{package}
\name{khtools-package}
\alias{khtools}
\alias{khtools-package}
\title{khtools: R package to make life in R a bit easier.}
\description{
This is a collection of miscellaneous wrappers for pipelining and processing genomic data. In addition there are utilities for quick interactive programming.
}
\author{
\strong{Maintainer}: Kevin Hadi \email{keh2019@med.cornell.edu}

}
