% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{kpng}
\alias{kpng}
\title{open png device with defaults}
\usage{
kpng(
  filename = "plot.png",
  height = 10,
  width = 10,
  h = height,
  w = width,
  cex = 1,
  title = NULL,
  byrow = TRUE,
  dim = NULL,
  cex.title = 1,
  oma.scale = 0,
  units = "in",
  oma.val = c(1, 1, 1, 1),
  useDingbats = FALSE,
  res = 300,
  pars = list(),
  ...
)
}
\description{
output to png device in default directory and keep open
}
