% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{len}
\alias{len}
\title{similar to length except gets nrows for those items that have dimensions}
\usage{
len(x, use.names = TRUE)
}
\arguments{
\item{x}{an object}
}
\value{
length or nrow of an object
}
\description{
figure out length or nrow of an object
}
