% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{match3}
\alias{match3}
\title{similar to setkey except a general use utility

very slow version of keying a la data.table
but for general/interactive use}
\usage{
match3(x, table, nomatch = NA_integer_)
}
\description{
similar to setkey except a general use utility

very slow version of keying a la data.table
but for general/interactive use
}
