% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{mstrsplit}
\alias{mstrsplit}
\title{make matrix out of stringsplitted character vector}
\usage{
mstrsplit(x, ...)
}
\value{
a matrix
}
\description{
split a vector by a delimiter,
fill in to make same length
bind into matrix
not as efficient as data.table::tstrsplit
but this is entirely base R
}
