% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{normpath}
\alias{normpath}
\title{normalize directory, but not basepath

get the absolute file directory without following the
base path link}
\usage{
normpath(p)
}
\arguments{
\item{str}{a path string}
}
\value{
a normalized path
}
\description{
normalize directory, but not basepath

get the absolute file directory without following the
base path link
}
