% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{parse.grl2}
\alias{parse.grl2}
\title{a robust version of parse.grl that is able to convert ranges with minus signs}
\usage{
parse.grl2(str, meta = NULL)
}
\value{
GRangesList
}
\description{

}
\author{
Kevin Hadi
}
