% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pg_mytheme}
\alias{pg_mytheme}
\title{print gg_mytheme output}
\usage{
pg_mytheme(..., print = TRUE)
}
\value{
A ggplot object
}
\description{
convenience wrapper around gg_mytheme
to automatically print the output
}
