% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pinch}
\alias{pinch}
\title{pinch a vector}
\usage{
pinch(x, fmin = 0.01, fmax = 0.99)
}
\value{
A vector
}
\description{
A convenience function to transform proportions.
Useful for beta regression (library(betareg))
Can be used on any numeric values to squeeze between
an interval.
}
