% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{ppng}
\alias{ppng}
\title{kevin's modification of ppng

height and width are specified in inches by default
resolution is specified as 300 by default}
\usage{
ppng(
  expr,
  filename = "plot.png",
  height = 10,
  width = 10,
  dim = NULL,
  cex = 1,
  title = NULL,
  h = height,
  w = width,
  cex.title = 1,
  oma.scale = 0,
  units = "in",
  res = 300,
  oma.val = c(1, 1, 1, 1),
  pars = list(),
  ...
)
}
\description{
kevin's modification of ppng

height and width are specified in inches by default
resolution is specified as 300 by default
}
