% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rm_mparen}
\alias{rm_mparen}
\title{utility function for removing multiple parantheses}
\usage{
rm_mparen(str)

rm_mparen(str)
}
\arguments{
\item{str}{a path string}
}
\value{
a string with multiple parentheses replaced with a single parenthesis

a string with multiple parentheses replaced with a single parenthesis
}
\description{
probably not necessary
}
