% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rn2col}
\alias{rn2col}
\title{alias for rownames_to_column

internal version that doesn't require library(tibble)}
\usage{
rn2col(.data, var = "rowname", keep.rownames = FALSE)
}
\arguments{
\item{.data}{a data frame/table}
}
\value{
a data frame/table with rownames from a column
}
\description{
alias for rownames_to_column

internal version that doesn't require library(tibble)
}
