% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rownames2}
\alias{rownames2}
\title{robust rownames

gives back character vector same number of rows of input regardless whether named or not}
\usage{
rownames2(x)
}
\arguments{
\item{str}{a path string}
}
\value{
a string with multiple parentheses replaced with a single parenthesis
}
\description{
robust rownames

gives back character vector same number of rows of input regardless whether named or not
}
