% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{rownames_to_column}
\alias{rownames_to_column}
\title{making column out of rownames}
\usage{
rownames_to_column(
  .data,
  var = "rowname",
  keep.rownames = FALSE,
  asdf = as.data.frame,
  as.data.frame = FALSE
)
}
\arguments{
\item{.data}{a data frame/table}
}
\value{
a data frame/table with the rownames as an additional column
}
\description{
internal version that doesn't require library(tibble)
}
