% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{setcolnames}
\alias{setcolnames}
\title{convenience function to set column names

alias of setColnames}
\usage{
setcolnames(object = nm, nm = NULL, pattern = NULL, replacement = "")
}
\arguments{
\item{object}{tabled object}

\item{nm}{names of the new columns}
}
\value{
colnamed object
}
\description{
convenience function to set column names

alias of setColnames
}
