% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{summ_glm}
\alias{summ_glm}
\title{tabular summary of glm model coefficients

Parse a tabular summary of a glm model}
\usage{
summ_glm(glm_mod, as.data.table = TRUE, ...)
}
\value{
A data.frame/data.table
}
\description{
tabular summary of glm model coefficients

Parse a tabular summary of a glm model
}
