% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{write_bed}
\alias{write_bed}
\title{write bed or bedpe into canonical formatted table

Write a table into  a bed oe bedpe formatted file
Ensures the header is commented out}
\usage{
write_bed(bed, outpath)
}
\value{
A data.table
}
\description{
write bed or bedpe into canonical formatted table

Write a table into  a bed oe bedpe formatted file
Ensures the header is commented out
}
